% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliasFromRunline.R
\name{aliasFromRunline}
\alias{aliasFromRunline}
\title{Derive the module Alias from a given runline}
\usage{
aliasFromRunline(line)
}
\arguments{
\item{line}{A line from a BioLockJ module run order}
}
\value{
the alias
}
\description{
Derive the module Alias from a given runline
}
