% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyShinyFeedback.R
\name{applyShinyFeedback}
\alias{applyShinyFeedback}
\title{apply shiny feedback}
\usage{
applyShinyFeedback(
  inputId,
  prop,
  feedbackId = inputId,
  moduleClass = NULL,
  projPath = projectDirPath(),
  input
)
}
\arguments{
\item{inputId}{the input object to pull a value from}

\item{prop}{a property object, such as an element of the list returned by BioLockR::propInfo()}

\item{feedbackId}{the input object to apply the feedback onto (by default, feedbackId=inputId)}

\item{moduleClass}{the class name of an associated module, used for the module arg to BioLockR::isValidProp}

\item{projPath}{use projectDirPath(); the path that relative paths are relative to.}

\item{input}{generally use: input=input}
}
\description{
apply shiny feedback
}
