% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_deployment.R
\name{detect_deployment}
\alias{detect_deployment}
\title{detect deployment}
\usage{
detect_deployment()
}
\value{
string, one of "remote", "virtual", or "local"
}
\description{
Determine if the the app is running on a local machine, on a virtual
machine (ie, from a docker container) or from a host such as shiny.io
}
