% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doIncludeProp.R
\name{doIncludeProp}
\alias{doIncludeProp}
\title{Include Property}
\usage{
doIncludeProp(property, value, default = NULL, input)
}
\arguments{
\item{property}{property name}

\item{value}{property value}

\item{default}{the default property for the value (if any)}

\item{input}{in general, input=input}
}
\value{
TRUE if the property should be included in the config file
}
\description{
Determine if a given property should be include in the config file.
}
