% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filePathChoices.R
\name{filePathChoices}
\alias{filePathChoices}
\title{Update File Path Choices}
\usage{
filePathChoices(existing, newPath = NULL)
}
\arguments{
\item{existing}{current string value given in properties}

\item{newPath}{optional, new path to add to list}
}
\value{
the new set of choices, essentially c(existing, newPath)
}
\description{
Update File Path Choices
}
