% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findExampleConfigs.R
\name{findExampleConfigs}
\alias{findExampleConfigs}
\title{Find Example Configs}
\usage{
findExampleConfigs(bljDir = dirname(dirname(BioLockR::get_BLJ_JAR())))
}
\arguments{
\item{bljDir}{}
}
\value{

}
\description{
Look in the greater BioLockJ folder that contains the currently configured BioLockJ.jar file.
Examples are expected to be under BioLockJ/templates/exampleProject/example.config,
where "BioLockJ" is the path passed in as \emph{bljDir}.
}
