% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModuleOverrideProp.R
\name{getModuleOverrideProp}
\alias{getModuleOverrideProp}
\title{Module Override Prop}
\usage{
getModuleOverrideProp(moduleAlias, propName)
}
\arguments{
\item{moduleAlias}{the alias of a BioLockJ module}

\item{propName}{a BioLockJ property}
}
\value{
the new property name
}
\description{
For a given module name/alias, and a given property, what would be the module-override property.
}
\examples{
getModuleOverrideProp("Part3", "script.numThreads")
#> [1] "Part3.numThreads"
}
