% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModuleRunLines.R
\name{getModuleRunLines}
\alias{getModuleRunLines}
\title{Module Run Lines}
\usage{
getModuleRunLines(moduleInfo)
}
\arguments{
\item{moduleInfo}{}
}
\value{
named vector
}
\description{
Given a lis of module objects, such as that returned by BioLockJ::moduleInfo(),
create a named vector linking the simple class name to the usage line.
}
