% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupPropsByCategory.R
\name{groupPropsByCategory}
\alias{groupPropsByCategory}
\title{Group Properties By Category}
\usage{
groupPropsByCategory(propInfo)
}
\arguments{
\item{propInfo}{}
}
\value{

}
\description{
The portion of a property name that comes before the first '.' is (generally) its category.
}
