% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_key_module_prop.R
\name{is_key_module_prop}
\alias{is_key_module_prop}
\title{Is this a key property}
\usage{
is_key_module_prop(prop, genPropInfo = BioLockR::propInfo())
}
\arguments{
\item{prop}{a property object, an element from the list returned by BioLockR::propInfo()}

\item{genPropInfo}{the result of BioLockR::propInfo(), or similar}
}
\value{
boolean
}
\description{
The logic for this mimics the logic used to separate properties in the user guide.
A property is a "key" property for a given module if the property is not also a general property OR
if the description is different from the general property description.
}
