% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isBadPrecheckResponse.R,
%   R/isGoodPrecheckResponse.R
\name{isBadPrecheckResponse}
\alias{isBadPrecheckResponse}
\alias{isGoodPrecheckResponse}
\title{Precheck Response}
\usage{
isBadPrecheckResponse(response)

isGoodPrecheckResponse(response)
}
\arguments{
\item{response}{the output of the precheck command to biolockj}
}
\value{
isBadPrecheckResponse: boolean, TRUE if the response is BAD

isGoodPrecheckResponse: boolean, TRUE if the response is GOOD
}
\description{
Determine if a response should overall be viewed as good (isGoodPrecheckResponse) or bad (isBadPrecheckResponse)
}
\section{Functions}{
\itemize{
\item \code{isBadPrecheckResponse}: check if the response is BAD

\item \code{isGoodPrecheckResponse}: check if the response is GOOD
}}

