% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isSharedProp.R
\name{isSharedProp}
\alias{isSharedProp}
\title{Shared Properties}
\usage{
isSharedProp(
  propMods = modulePerProp(),
  propName,
  includeMods = unique(unlist(propMods))
)
}
\arguments{
\item{propMods}{output of modulePerProp}

\item{propName}{property in question}

\item{includeMods}{the modules in the current pipeline}
}
\value{
boolean
}
\description{
Determines if the given property is used by more than one of the modules in the pipeline.
}
