% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isValidAlias.R
\name{isValidAlias}
\alias{isValidAlias}
\title{Test if a string is a valid module alias}
\usage{
isValidAlias(alias, existingAlia = c())
}
\arguments{
\item{alias}{the candidate alias}

\item{existingAlia}{character vector of the aliases currently in use}
}
\value{
boolean
}
\description{
Test if an alias is usable for a BioLockJ pipeline module.
}
\details{
If false, a message is printed to stderr with explaination.
}
\examples{
isValidAlias("foo")
#> FALSE
isValidAlias("Foo")
#> TRUE
isValidAlias("Foo", c("Foo", "Bar", "Baz"))
#' #> FALSE
}
