% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isWritableValue.R
\name{isWritableValue}
\alias{isWritableValue}
\title{Is Writable Value}
\usage{
isWritableValue(value)
}
\arguments{
\item{value}{}
}
\value{
boolean
}
\description{
Test that a value is something a method can work with.
}
\examples{
isWritableValue(NULL)
#> [1] FALSE
isWritableValue(NA)
#> [1] FALSE
isWritableValue("apple")
#> [1] TRUE
isWritableValue( -5 )
#> [1] TRUE
isWritableValue( c(1,2,3))
#> [1] TRUE
}
