% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRunLine.R
\name{makeRunLine}
\alias{makeRunLine}
\title{Assemble the run line for a module}
\usage{
makeRunLine(moduleName, moduleRunLines, alias = "")
}
\arguments{
\item{moduleName}{the simple name of a module class}

\item{moduleRunLines}{named vector linking the module name to the full classpath preceeded by keyword}

\item{alias}{optional alias}
}
\value{
The line to include in a BioLockJ module run order
}
\description{
Assemble the run line for a module
}
