% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdownTextToHTML.R
\name{markdownTextToHTML}
\alias{markdownTextToHTML}
\title{markdownToHTML}
\usage{
markdownTextToHTML(text)
}
\arguments{
\item{text}{passed to the text argument in markdown::markdownToHTML}
}
\value{
html
}
\description{
A modified form of the markdown::markdownToHTML function.
}
