% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_prop_UI_name.R
\name{module_prop_UI_name}
\alias{module_prop_UI_name}
\title{Name UI elements for BioLockJ Properties}
\usage{
module_prop_UI_name(propName, module)
}
\arguments{
\item{propName}{the name of a BioLockJ property, such as "script.numThreads"}
}
\value{
a string to use for an associated UI element
}
\description{
Name UI elements for BioLockJ Properties
}
