% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_userguide_url.R
\name{module_userguide_url}
\alias{module_userguide_url}
\title{Create a user guide link for a module.}
\usage{
module_userguide_url(
  class,
  baseUrl = "https://biolockj-dev-team.github.io/BioLockJ/GENERATED/"
)
}
\arguments{
\item{class}{class path for a BioLockJ module}

\item{baseUrl}{the base url for the user guide; see details}
}
\value{
url as a string
}
\description{
Create a user guide link for a module.
}
