% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namedDefaultVals.R
\name{namedDefaultVals}
\alias{namedDefaultVals}
\title{Named Default Values}
\usage{
namedDefaultVals(propInfo)
}
\arguments{
\item{propInfo}{the result of BioLockJ::propInfo() or a similar list of property atrribute lists.}
}
\value{
a named character vector where the names are the property names and the values are the property values
}
\description{
Given a list of property infos, get a named vector of default values.
}
\details{
Only the properties that have default values are included in the output.
}
