% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propInfoId.R
\name{propInfoId}
\alias{propInfoId}
\title{Title}
\usage{
propInfoId(propUiName)
}
\arguments{
\item{propUiName}{often the output of propUiName or module_prop_UI_name}
}
\value{
an id
}
\description{
Title
}
