% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propInfoSansSpecials.R
\name{propInfoSansSpecials}
\alias{propInfoSansSpecials}
\title{Limit Prop Info for GUI}
\usage{
propInfoSansSpecials()
}
\value{
similar to BioLockR::PropInfo()
}
\description{
Get the properties from BioLockJ, but remove the properties that are not desired.
}
