% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propsInfoForType.R
\name{propsInfoForType}
\alias{propsInfoForType}
\title{Property Info For Type}
\usage{
propsInfoForType(propInfo, type)
}
\arguments{
\item{propInfo}{named list of property objects}

\item{type}{the type of property to keep}
}
\value{
a named list of property objects, a subset of propInfo.
}
\description{
Get a subset of the propInfo() value for only one category of property.
}
