% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propClearBtn.R, R/propDirChooserId.R,
%   R/propFileChooserId.R, R/propInfoId.R, R/propOverrideBtnId.R,
%   R/propRmOverrideBtnId.R, R/propSelectFromId.R, R/propShowId.R,
%   R/propUiName.R
\name{propClearBtn}
\alias{propClearBtn}
\alias{propDirChooserId}
\alias{propFileChooserId}
\alias{propInfoId}
\alias{propOverrideBtnId}
\alias{propRmOverrideBtnId}
\alias{propSelectFromId}
\alias{propShowId}
\alias{propUiName}
\title{Name UI elements for BioLockJ Properties}
\usage{
propClearBtn(propName)

propDirChooserId(propName)

propFileChooserId(propName)

propInfoId(propName)

propOverrideBtnId(propName, module)

propRmOverrideBtnId(propName, module)

propSelectFromId(propName)

propShowId(propName)

propUiName(propName)
}
\arguments{
\item{propName}{the name of a BioLockJ property, such as "script.numThreads"}

\item{module}{the module alias}
}
\value{
a string to use for an associated UI element
}
\description{
Name UI elements for BioLockJ Properties
}
\section{Functions}{
\itemize{
\item \code{propClearBtn}: name for button to clear the value of a property

\item \code{propDirChooserId}: name for a directory chooser ui element

\item \code{propFileChooserId}: name for a file chooser ui element

\item \code{propInfoId}: 

\item \code{propOverrideBtnId}: name for a button to create a module-instance-override form of the property

\item \code{propRmOverrideBtnId}: name for a button to remove a module-instance-override form of the property

\item \code{propSelectFromId}: name for selector tool as part of a ui element

\item \code{propShowId}: name for ui element that shows the property value
}}

