% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderPropUi.R
\name{renderPropUi}
\alias{renderPropUi}
\title{General BioLockJ Properties UI}
\usage{
renderPropUi(propName, prop, default, value, defaults)
}
\arguments{
\item{propName}{a property name, such script.numThreads}

\item{prop}{a property object, which is a list that contains type, description, etc}

\item{default}{the current default value}

\item{value}{the current value}

\item{defaults}{the list of lists of representation of defaults}
}
\value{

}
\description{
As of this writing, the Properties class in BioLockJ has these KNOWN_TYPES:
KNOWN_TYPES = {STRING_TYPE, BOOLEAN_TYPE, FILE_PATH, EXE_PATH, LIST_TYPE, FILE_PATH_LIST, INTEGER_TYPE, NUMERTIC_TYPE};
KNOWN_TYPES = {"string", "boolean", "file path", "executable", "list", "list of file paths", "integer", "numeric"};
}
