% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_biolockj.R
\name{shiny_biolockj}
\alias{shiny_biolockj}
\title{shiny biolockj}
\usage{
shiny_biolockj(...)
}
\arguments{
\item{...}{parameters passed to shinyApp(ui, server, ...)}
}
\value{
no return value
}
\description{
This is the essential launch command for the shiny app.
}
