% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateSharedPropUi.R
\name{updateSharedPropUi}
\alias{updateSharedPropUi}
\title{Update the values in the ui controls for shared properties}
\usage{
updateSharedPropUi(session, prop, moduleId, value)
}
\arguments{
\item{session}{session=session}

\item{prop}{the property object, must have a $type}

\item{moduleId}{the module whose ui must be updated}

\item{value}{the new value to apply}
}
\description{
Update the values in the ui controls for shared properties
}
\seealso{
renderPropUi, buildFilePathPropUI, buildFileListPropUI
}
