% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeConfigProp.R
\name{writeConfigProp}
\alias{writeConfigProp}
\title{Format a property for BioLockJ}
\usage{
writeConfigProp(
  propName,
  propVal = NULL,
  propType = "string",
  projectDir = "",
  useRelPath = FALSE
)
}
\arguments{
\item{propName}{the name of the property}

\item{propVal}{the value for the property}

\item{propType}{the type of property}

\item{projectDir}{the project directory (the folder where the config file lives), only used if \emph{useRelPath=TRUE}}

\item{useRelPath}{boolean, should file paths be written as relative to the projectDir}
}
\value{

}
\description{
Write a property in the format suitable for a BioLockJ property.
}
