% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeFilePath.R, R/writeFilePathList.R,
%   R/writeFullPath.R, R/writeRelPath.R
\name{writeFilePath}
\alias{writeFilePath}
\alias{writeFilePathList}
\alias{writeFullPath}
\alias{writeRelPath}
\title{File Paths for BioLockJ}
\usage{
writeFilePath(path, projectDir, useRelPath)

writeFilePathList(pathList, projectDir, useRelPath)

writeFullPath(path, projectDir)

writeRelPath(path, projectDir)
}
\arguments{
\item{path}{a file path (written as either from root or starting with ./ or ../)}

\item{projectDir}{the path that the paths should be relative to}

\item{useRelPath}{if TRUE, relative paths are written, if false, then full paths are determined.}

\item{pathList}{the value of a list property, probably represented as a single string}
}
\value{
a string representing the file path
}
\description{
Write file paths for a BioLockJ config file.
}
\details{
Paths can be passed in as EITHER relative to root or relative to project dir (ie, starting with './' or '../').
\emph{useRelPath} dictacts how the output is written, not how the input is interpreted.
}
\section{Functions}{
\itemize{
\item \code{writeFilePathList}: use writeFilePath for each element of a list.

\item \code{writeFullPath}: write the path from root.

\item \code{writeRelPath}: write the path from the project directory (ie, ./ or ../) if possible.
}}

\examples{
writeFilePath("/Users/me/my/path/to/file.txt", "/Users/me/my/path", FALSE)
#> [1] "/Users/me/my/path/to/file.txt"
writeFilePath("/Users/me/my/path/to/file.txt", "/Users/me/my/path", TRUE)
#> [1] "/Users/me/my/path/to/file.txt"
}
