% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyPropOwnership.R
\name{applyPropOwnership}
\alias{applyPropOwnership}
\title{Title}
\usage{
applyPropOwnership(modules, genProps)
}
\arguments{
\item{modules}{list of module info, elements should be named, generally similar what is returned by BioLockR::moduleInfo()}

\item{genProps}{vector of general property names}
}
\value{
a list similar to \emph{modules}, but where the properties objects have added elements "override" and "ownership";
override is the module-specific override property for this this module and property and
ownership is one of "general", "shared", or "single"
}
\description{
Title
}
