% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biolockj_server.R
\name{biolockj_server}
\alias{biolockj_server}
\title{BioLockJ server.R}
\usage{
biolockj_server(input, output, session)
}
\arguments{
\item{input}{}

\item{output}{}

\item{session}{}
}
\value{
server function
}
\description{
BioLockJ server.R
}
