% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFileListPropObservers.R,
%   R/buildFileListPropUI.R, R/buildFilePathPropObservers.R,
%   R/buildFilePathPropUI.R
\name{buildFileListPropObservers}
\alias{buildFileListPropObservers}
\alias{buildFileListPropUI}
\alias{buildFilePathPropObservers}
\alias{buildFilePathPropUI}
\title{Build UI for File Path Properties}
\usage{
buildFileListPropObservers(
  session,
  input,
  output,
  propName,
  myVolumesNow,
  values
)

buildFileListPropUI(propName, value, default, moduleId = NULL)

buildFilePathPropObservers(
  session,
  input,
  output,
  propName,
  myVolumesNow,
  values
)

buildFilePathPropUI(propName, value, default, moduleId = NULL)
}
\arguments{
\item{session}{usually session=session}

\item{input}{usually input=input}

\item{output}{usually output=output}

\item{propName}{a string, the name of the property}

\item{myVolumesNow}{the volumes object to pass to shinyFiles operations}

\item{values}{the list of lists reactive values object that represents the current configuration}

\item{value}{in the case of a list property, the value of the property, most likely a single string, possibly comma-separated list}

\item{default}{value shown as a placeholder if no value is present, or value is removed}

\item{moduleId}{the module id if the property ui is associated with a module.}
}
\value{


ui object
}
\description{
Build UI for File Path Properties
}
\section{Functions}{
\itemize{
\item \code{buildFileListPropObservers}: Build/refresh the backend of the UI for a BioLockJ file list property

\item \code{buildFileListPropUI}: Build the UI for a BioLockJ file list property

\item \code{buildFilePathPropObservers}: Build/refresh the backend of the UI for a BioLockJ file property

\item \code{buildFilePathPropUI}: Build the UI for a BioLockJ file property
}}

