% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classFromRunline.R
\name{classFromRunline}
\alias{classFromRunline}
\title{Derive the module class from a given runline}
\usage{
classFromRunline(line)
}
\arguments{
\item{line}{A line from a BioLockJ module run order}
}
\value{
the class of the module specified by the runline
}
\description{
Derive the module class from a given runline
}
\examples{
line = "#BioModule path.to.class.for.Dostuff AS MyStuff"
classFromRunline( line )
#> "path.to.class.for.Dostuff"

}
\seealso{
aliasFromRunline
}
