% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isWritableValue.R
\name{isWritableValue}
\alias{isWritableValue}
\title{Is Writable Value}
\usage{
isWritableValue(value)
}
\arguments{
\item{value}{The value to test}
}
\value{
boolean
}
\description{
Test that a value is something a method can work with.
}
