% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRunLine.R
\name{makeRunLine}
\alias{makeRunLine}
\title{Assemble the run line for a module}
\usage{
makeRunLine(moduleName, moduleRunLines, alias = "")
}
\arguments{
\item{moduleName}{the simple name of a module class}

\item{moduleRunLines}{named vector linking the module name to the full classpath preceeded by keyword}

\item{alias}{optional alias}
}
\value{

}
\description{
Assemble the run line for a module
}
\examples{
avail = c(dostuff="#BioModule path/to/class/for/dostuff",
thing="#BioModule path/to/class/for/thing")
makeRunLine("dostuff", avail, "MyStuff")
#> [1] "#BioModule path/to/class/for/dostuff AS MyStuff"
}
