% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulePerProp.R
\name{modulePerProp}
\alias{modulePerProp}
\title{Module Per Prop}
\usage{
modulePerProp(moduleInfo = BioLockR::moduleInfo())
}
\arguments{
\item{moduleInfo}{the lis returned by BioLockJ::moduleInfo(), or a similar object}
}
\value{
a list of properties, each element is a vector of module names giving the modules that use that property
}
\description{
List the modules organized by property name.  Thus, get a list of all modules that use a given property.
}
