% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderDefaultPropFiles.R
\name{orderDefaultPropFiles}
\alias{orderDefaultPropFiles}
\title{Order Default Props Files}
\usage{
orderDefaultPropFiles(start, chain)
}
\arguments{
\item{start}{a file name (basename) that correstponds to one of the names of the chain list}

\item{chain}{a named list whose names are file names (basename) and whose elements are
each the vecter of file names (full file path) that the named file lists as default props.}
}
\value{
a list of:
missing: full file paths whose basename is not in names(chain)
dangling: The name of the file that lead to the file path(s) that are missing
chained: a vector of basenames, ending with 'start'.  This is the order in which to load the available default props files.
}
\description{
Model default props inter-dependences
}
