% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propClearBtn.R, R/propDirChooserId.R,
%   R/propEditBtnId.R, R/propFileChooserId.R, R/propInfoId.R,
%   R/propModuleFlipPanel.R, R/propOverrideBtnId.R, R/propRmOverrideBtnId.R,
%   R/propShowId.R, R/propUiName.R
\name{propClearBtn}
\alias{propClearBtn}
\alias{propDirChooserId}
\alias{propEditBtnId}
\alias{propFileChooserId}
\alias{propInfoId}
\alias{propModuleFlipPanel}
\alias{propOverrideBtnId}
\alias{propRmOverrideBtnId}
\alias{propShowId}
\alias{propUiName}
\title{Name UI elements for BioLockJ Properties}
\usage{
propClearBtn(propName, moduleId = NULL)

propDirChooserId(propName, moduleId = NULL)

propEditBtnId(propName, moduleId)

propFileChooserId(propName, moduleId = NULL)

propInfoId(propName, moduleId = NULL)

propModuleFlipPanel(propName, moduleId)

propOverrideBtnId(propName, moduleId)

propRmOverrideBtnId(propName, moduleId)

propShowId(propName, moduleId)

propUiName(propName, moduleId = NULL)
}
\arguments{
\item{propName}{the name of a BioLockJ property, such as "script.numThreads"}

\item{moduleId}{the module id}
}
\value{
an id

a string to use for an associated UI element
}
\description{
Name UI elements for BioLockJ Properties
}
\section{Functions}{
\itemize{
\item \code{propClearBtn}: name for button to clear the value of a property

\item \code{propDirChooserId}: name for a directory chooser ui element

\item \code{propEditBtnId}: name for a button to switch focus to a ui where you can edit a displayed property

\item \code{propFileChooserId}: name for a file chooser ui element

\item \code{propInfoId}: the ui id for the info icon about defaults

\item \code{propModuleFlipPanel}: name for a tabpanel within a module section, flips between panel for active ui and panel for viewing the proprety that is set elsehwere.

\item \code{propOverrideBtnId}: name for a button to create a module-instance-override form of the property

\item \code{propRmOverrideBtnId}: name for a button to remove a module-instance-override form of the property

\item \code{propShowId}: name for ui element that shows the property value
}}

