% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderPropUi.R
\name{renderPropUi}
\alias{renderPropUi}
\title{General BioLockJ Properties UI}
\usage{
renderPropUi(
  propName,
  prop,
  value,
  defaults,
  ownership = "general",
  moduleId = NULL,
  trailingUiFun = function() {     tagList(hr()) }
)
}
\arguments{
\item{propName}{a property name, such script.numThreads}

\item{prop}{a property object, which is a list that contains type, description, etc}

\item{value}{a string, the current value}

\item{defaults}{the list of lists of representation of defaults}

\item{ownership}{the ownership category for the proprety, one of c("general", "shared", "owned", "override")}

\item{moduleId}{if ownership is not "general", which module is the ownership referenceing}

\item{trailingUiFun}{a function that takes no args and returns a tagList object}
}
\value{
ui object list
}
\description{
Create the ui for an individual biolockj property.
}
\details{
As of this writing, the Properties class in BioLockJ has these KNOWN_TYPES:
KNOWN_TYPES = {STRING_TYPE, BOOLEAN_TYPE, FILE_PATH, EXE_PATH, LIST_TYPE, FILE_PATH_LIST, INTEGER_TYPE, NUMERTIC_TYPE};
KNOWN_TYPES = {"string", "boolean", "file path", "executable", "list", "list of file paths", "integer", "numeric"};
}
\seealso{
updateSharedPropUi
}
